Move authentication options to auth file
